Refactor en paquet sms_api
