+++ docs/mise-a-jour.md
@@ -3,6 +3,7 @@
 Cette page recense les évolutions majeures de l'application. Elle doit être mise à jour à chaque merge sur la branche `main`.
 
 ## Historique
+- **21 juillet 2025** : refonte en modules (`sms_api`) et simplification de `sms_http_api.py`.
 - **21 juillet 2025** : ajout d'un champ pour saisir la clef `X-API-KEY` dans la page "Test SMS".
 - **21 juillet 2025** : ajout de l'option `--api-key` pour sécuriser l'envoi de SMS via l'en-tête `X-API-KEY`.
 - **20 juillet 2025** : ajout d'une pastille indiquant le nombre de SMS reçus dans le menu.
