+++ sms_api/server.py
@@ -1,0 +1,22 @@
+from http.server import HTTPServer
+
+
+class SMSHTTPServer(HTTPServer):
+    def __init__(
+        self,
+        server_address,
+        handler_class,
+        modem_url,
+        username,
+        password,
+        db_path,
+        api_key=None,
+    ):
+        super().__init__(server_address, handler_class)
+        self.modem_url = modem_url
+        self.username = username
+        self.password = password
+        self.db_path = db_path
+        self.api_key = api_key
+
+__all__ = ["SMSHTTPServer"]
